% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMotifBed.R
\name{mk.motif.bed}
\alias{mk.motif.bed}
\title{Create bed file of motif you are interested in}
\arguments{
\item{Motif}{Motif you are interested in}

\item{Genome}{Input fasta file of genome you are using}
}
\value{
Dataframe of input motif in bed file format
}
\description{
/code{mk.motif.bed} create bed file of motif you are interested in
}
